Sterted ECC crypto

+++ SafeCamera/src/main/java/com/fenritz/safecam/util/Helpers.java
@@ -161,8 +161,10 @@ public class Helpers {
 		File dir = new File(Helpers.getHomeDir(context) + "/" + ".tmp");
 		if (dir.isDirectory()) {
 			String[] children = dir.list();
-			for (int i = 0; i < children.length; i++) {
-				new File(dir, children[i]).delete();
+			if(children != null) {
+				for (int i = 0; i < children.length; i++) {
+					new File(dir, children[i]).delete();
+				}
 			}
 		}
 	}
